requirements.txt
djangorestframework==3.15.2
django-filter==25.1
gunicorn==23.0.0

psycopg2-binary==2.9.10

djangorestframework-jsonapi==7.1.0
dj-database-url==2.3.0
drf_spectacular==0.28.0
drf-spectacular-jsonapi==0.5.2
